Class renaming and slight cleaning
