--- BrAPI_to_BioSamples/BioSamplesData/getBioSample.sh
@@ -1,1 +1,0 @@
-curl -X GET -H "Accept: application/hal+json" "http://byod.psblocal:8081/biosamples/beta/samples/$1"
