--- BrAPI_to_BioSamples/BioSamplesData/submitBioSample.sh
@@ -1,1 +1,0 @@
-cat "$1/$2.json" | curl -X PUT -H "Content-Type: application/json" -d @- "http://byod.psblocal:8081/biosamples/beta/samples/$2"
